:wrench: add check for .zshrc

+++ set_zshrc.sh
@@ -1,3 +1,8 @@
-#!/bin/sh
-exec cp ~/.zshrc ~/.zshrc.backup
+#!/bin/bash
+ZSHRC="~/.zshrc"
+if [ -e $ZSHRC ]; then
+  echo ".zshrc found. Creating backup..."
+  exec cp ~/.zshrc ~/.zshrc.backup
+fi
 exec cp .zshrc ~/.zshrc 
+echo "Successfully copied .zshrc"
